Moved Command-related data structures into core
